refactor: break out loginterceptor to a separate file
