+++ coreweave/client_test.go
@@ -1,163 +1,14 @@
 package coreweave_test
 
 import (
-	"bytes"
-	"context"
 	"errors"
 	"testing"
 
-	"buf.build/gen/go/coreweave/cks/connectrpc/go/coreweave/cks/v1beta1/cksv1beta1connect"
-	cksv1beta1 "buf.build/gen/go/coreweave/cks/protocolbuffers/go/coreweave/cks/v1beta1"
 	"connectrpc.com/connect"
 	"github.com/coreweave/terraform-provider-coreweave/coreweave"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
-	"github.com/hashicorp/terraform-plugin-log/tflogtest"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
-
-func exampleClusterRequest() *cksv1beta1.CreateClusterRequest {
-	return &cksv1beta1.CreateClusterRequest{
-		Name:   "test-cluster",
-		Zone:   "US-EAST-04A",
-		VpcId:  "8e727c0d-5527-416b-9a80-c498f4802d15",
-		Public: true,
-	}
-}
-
-func exampleClusterResponse() *cksv1beta1.CreateClusterResponse {
-	return &cksv1beta1.CreateClusterResponse{
-		Cluster: &cksv1beta1.Cluster{
-			Id:     "02762697-916a-4920-940b-5909780f358e",
-			Name:   "test-cluster",
-			Zone:   "US-EAST-04A",
-			Status: cksv1beta1.Cluster_STATUS_CREATING,
-			VpcId:  "8e727c0d-5527-416b-9a80-c498f4802d15",
-			Public: true,
-		},
-	}
-}
-
-func TestTFLogInterceptor(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name       string
-		endpoint   string
-		req        *connect.Request[cksv1beta1.CreateClusterRequest]
-		returnResp *connect.Response[cksv1beta1.CreateClusterResponse]
-		returnErr  error
-		assertion  func(t *testing.T, logEntries []map[string]any)
-	}{
-		{
-			name:       "Basic successful response",
-			endpoint:   "https://api.coreweave.com",
-			req:        connect.NewRequest(exampleClusterRequest()),
-			returnResp: connect.NewResponse(exampleClusterResponse()),
-			assertion: func(t *testing.T, logEntries []map[string]any) {
-				t.Helper()
-				assert.Len(t, logEntries, 2)
-				respEntry := logEntries[1]
-				t.Run("request matches", func(t *testing.T) {
-					t.Parallel()
-					reqEntry := logEntries[0]
-					assert.Contains(t, reqEntry["@message"], "sending API request")
-					assert.Equal(t, "connect://api.coreweave.com", reqEntry["peer"])
-					assert.Equal(t, "unary", reqEntry["streamType"])
-					assert.Equal(t, "/coreweave.cks.v1beta1.ClusterService/CreateCluster", reqEntry["procedure"])
-					assert.NotEmpty(t, reqEntry["payload"])
-
-					referenceReq := exampleClusterRequest()
-					assert.Contains(t, reqEntry["payload"], referenceReq.Name)
-					assert.Contains(t, reqEntry["payload"], referenceReq.Zone)
-					assert.Contains(t, reqEntry["payload"], referenceReq.VpcId)
-
-					assert.NotContains(t, reqEntry, "error")
-				})
-				t.Run("response matches", func(t *testing.T) {
-					t.Parallel()
-					assert.Contains(t, respEntry["@message"], "received API response")
-					assert.Equal(t, "connect://api.coreweave.com", respEntry["peer"])
-					assert.Equal(t, "unary", respEntry["streamType"])
-					assert.Equal(t, "/coreweave.cks.v1beta1.ClusterService/CreateCluster", respEntry["procedure"])
-					assert.NotEmpty(t, respEntry["payload"])
-					assert.NotContains(t, respEntry, "error")
-
-					referenceResp := exampleClusterResponse()
-					assert.Contains(t, respEntry["payload"], referenceResp.Cluster.Id)
-					assert.Contains(t, respEntry["payload"], referenceResp.Cluster.Name)
-					assert.Contains(t, respEntry["payload"], referenceResp.Cluster.VpcId)
-					assert.Contains(t, respEntry["payload"], referenceResp.Cluster.Status.String())
-				})
-			},
-		},
-		{
-			name:       "Basic error response",
-			endpoint:   "https://api.coreweave.com",
-			req:        connect.NewRequest(exampleClusterRequest()),
-			returnResp: nil,
-			returnErr:  connect.NewError(connect.CodeInternal, errors.New("Internal server error")),
-			assertion: func(t *testing.T, logEntries []map[string]any) {
-				t.Helper()
-				t.Run("request matches", func(t *testing.T) {
-					t.Parallel()
-					reqEntry := logEntries[0]
-					assert.Contains(t, reqEntry["@message"], "sending API request")
-					assert.Equal(t, "connect://api.coreweave.com", reqEntry["peer"])
-					assert.Equal(t, "unary", reqEntry["streamType"])
-					assert.Equal(t, "/coreweave.cks.v1beta1.ClusterService/CreateCluster", reqEntry["procedure"])
-					assert.NotEmpty(t, reqEntry["payload"])
-					assert.NotContains(t, reqEntry, "error")
-				})
-				t.Run("resp shows error", func(t *testing.T) {
-					t.Parallel()
-					respEntry := logEntries[1]
-					assert.Contains(t, respEntry["@message"], "got nil or invalid API response")
-					assert.Equal(t, "connect://api.coreweave.com", respEntry["peer"])
-					assert.Equal(t, "unary", respEntry["streamType"])
-					assert.Equal(t, "/coreweave.cks.v1beta1.ClusterService/CreateCluster", respEntry["procedure"])
-					assert.NotContains(t, respEntry, "payload")
-					assert.Contains(t, respEntry, "error")
-					assert.Contains(t, respEntry["error"], "Internal server error")
-				})
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			var logbuf bytes.Buffer
-			ctx := tflogtest.RootLogger(t.Context(), &logbuf)
-
-			interceptor := coreweave.TFLogInterceptor()
-			c := cksv1beta1connect.NewClusterServiceClient(nil, tt.endpoint, connect.WithInterceptors(interceptor, connect.UnaryInterceptorFunc(func(next connect.UnaryFunc) connect.UnaryFunc {
-				return func(ctx context.Context, req connect.AnyRequest) (connect.AnyResponse, error) {
-					// This short-circuits the client to return the desired response and error for testing without calling the next func.
-					return tt.returnResp, tt.returnErr
-				}
-			})))
-			resp, err := c.CreateCluster(ctx, tt.req)
-			logBytes := logbuf.Bytes() // Capture the log bytes so we can copy them as necessary.
-
-			assert.Equal(t, tt.returnErr, err)
-			assert.Equal(t, tt.returnResp, resp)
-
-			logEntries, err := tflogtest.MultilineJSONDecode(bytes.NewBuffer(logBytes))
-			require.NoError(t, err)
-
-			for i, entry := range logEntries {
-				if payload, ok := entry["payload"]; ok {
-					assert.NotContains(t, payload, "\n", "payload of msg %d should be single-line in logs", i)
-				}
-			}
-
-			if tt.assertion != nil {
-				tt.assertion(t, logEntries)
-			}
-		})
-	}
-}
 
 func TestHandleAPIError(t *testing.T) {
 	t.Parallel()
